+++ src/cli/Microsoft.DotNet.UpgradeAssistant.Cli/Commands/Analyze/ConsoleAnalyze.cs
@@ -3,8 +3,10 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.DotNet.UpgradeAssistant.Analysis;
 using Microsoft.Extensions.Logging;
 
 namespace Microsoft.DotNet.UpgradeAssistant.Cli
@@ -15,31 +17,39 @@ namespace Microsoft.DotNet.UpgradeAssistant.Cli
         private readonly IUpgradeStateManager _stateManager;
         private readonly ILogger<ConsoleAnalyze> _logger;
         private readonly IEnumerable<IAnalyzeResultProvider> _providers;
+        private readonly IAnalyzeResultWriter _writer;
 
         public ConsoleAnalyze(
             IEnumerable<IAnalyzeResultProvider> analysisProviders,
             IUpgradeContextFactory contextFactory,
             IUpgradeStateManager stateManager,
+            IAnalyzeResultWriter writer,
             ILogger<ConsoleAnalyze> logger)
         {
             _providers = analysisProviders;
+            _writer = writer;
             _contextFactory = contextFactory ?? throw new ArgumentNullException(nameof(contextFactory));
             _stateManager = stateManager ?? throw new ArgumentNullException(nameof(stateManager));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         }
 
-        [ObsoleteAttribute("This property is WIP, expect changes in this area.", false)]
         public async Task RunAsync(CancellationToken token)
         {
-            _logger.LogWarning("The Analyze command feature is still under development, expect things to be not fully functional at the moment");
             using var context = await _contextFactory.CreateContext(token);
 
             await _stateManager.LoadStateAsync(context, token);
             var analzyerContext = new AnalyzeContext(context);
+            var analyzeResultMap = new List<AnalyzeResultDefinition>();
             foreach (var provider in _providers)
             {
-                await provider.AnalyzeAsync(analzyerContext, token);
+                analyzeResultMap.Add(new()
+                {
+                    AnalysisTypeName = provider.AnalysisTypeName,
+                    AnalysisResults = await provider.AnalyzeAsync(analzyerContext, token),
+                });
             }
+
+            await _writer.WriteAsync(analyzeResultMap.ToAsyncEnumerable(), token).ConfigureAwait(false);
         }
     }
 }