+++ src/cli/Microsoft.DotNet.UpgradeAssistant.Cli/Commands/Analyze/ConsoleAnalyzeCommand.cs
@@ -13,8 +13,8 @@ namespace Microsoft.DotNet.UpgradeAssistant.Cli
         public ConsoleAnalyzeCommand()
             : base("analyze")
         {
-            IsHidden = true;
-            Handler = CommandHandler.Create<ParseResult, UpgradeOptions, CancellationToken>((result, options, token) =>
+            IsHidden = !FeatureFlags.IsRequested("ANALYZE");
+            Handler = CommandHandler.Create<ParseResult, CommandOptions, CancellationToken>((result, options, token) =>
                 Host.CreateDefaultBuilder()
                     .UseConsoleUpgradeAssistant<ConsoleAnalyze>(options, result)
                     .RunUpgradeAssistantAsync(token));