+++ src/common/Microsoft.DotNet.UpgradeAssistant.Abstractions/IAnalyzeResultProvider.cs
@@ -2,14 +2,16 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.DotNet.UpgradeAssistant
+namespace Microsoft.DotNet.UpgradeAssistant.Analysis
 {
     public interface IAnalyzeResultProvider
     {
-        [ObsoleteAttribute("This property is WIP, expect changes in this area.", false)]
-        Task AnalyzeAsync(AnalyzeContext analysis, CancellationToken token);
+        string AnalysisTypeName { get; }
+
+        Task<IAsyncEnumerable<AnalyzeResult>> AnalyzeAsync(AnalyzeContext analysis, CancellationToken token);
     }
 }