+++ src/common/Microsoft.DotNet.UpgradeAssistant.Abstractions/IAnalyzeResultWriter.cs
@@ -1,0 +1,16 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.DotNet.UpgradeAssistant.Analysis
+{
+    public interface IAnalyzeResultWriter
+    {
+        Task WriteAsync(IAsyncEnumerable<AnalyzeResultDefinition> results, CancellationToken token);
+    }
+}