+++ src/components/Microsoft.DotNet.UpgradeAssistant.Analysis/AnalysisExtensions.cs
@@ -1,0 +1,23 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using Microsoft.DotNet.UpgradeAssistant.Analysis;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Microsoft.DotNet.UpgradeAssistant
+{
+    public static class AnalysisExtensions
+    {
+        public static void AddAnalysis(this IServiceCollection services)
+        {
+            if (services is null)
+            {
+                throw new ArgumentNullException(nameof(services));
+            }
+
+            services.AddTransient<ISerializer, JsonSerializer>();
+            services.AddTransient<IAnalyzeResultWriter, AnalyzeResultWriter>();
+        }
+    }
+}