+++ src/extensions/default/analyzers/Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CodeFixes/HttpContextIsDebuggingEnabledCodeFixer.cs
@@ -15,7 +15,6 @@ using Microsoft.DotNet.UpgradeAssistant.Extensions.Default.Analyzers;
 
 namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CodeFixes
 {
-    [ApplicableComponents(ProjectComponents.AspNetCore)]
     [ExportCodeFixProvider(LanguageNames.CSharp, Name = "UA0006 CodeFix Provider")]
     public class HttpContextIsDebuggingEnabledCodeFixer : CodeFixProvider
     {